requirements.txt
azure-functions
fastapi
python-dotenv
pymongo
uvicorn
python-jose[cryptography]
passlib[bcrypt]
python-multipart
requests
phonenumbers
pydantic[email]